requirements.txt
numpy==1.19.0
opencv-python==4.2.0.32
pyserial==3.4
python-decouple==3.3